Simplify code: Leverage cubin/fmha.py ABI compatibility
